refactor: limit general getNodeBySubcontext to instance getFacetNode
