+++ client/lib/CategoryStratum/getFacetNode.js
