+++ client/lib/CategoryStratum/index.js
@@ -62,7 +62,7 @@ Object.assign(proto, Stratum.prototype)
 proto.enableFaceting = require('./enableFaceting')
 proto.filter = require('./filter')
 proto.getBasisForSubstratum = require('./getBasisForSubstratum')
-proto.getNodeBySubcontext = require('./getNodeBySubcontext')
+proto.getFacetNode = require('./getFacetNode')
 proto.load = require('./load')
 proto.prune = require('./prune')
 proto.refreshNodeSizes = require('./refreshNodeSizes')
