--- client/lib/Stratum/getNodeBySubcontext.js
@@ -1,7 +1,0 @@
-module.exports = function () {
-  // @Stratum:getNodeBySubcontext()
-  //
-  // A placeholder that ensures subclasses implement the method.
-  //
-  throw new Error('Subclass must implement the getNodeBySubcontext method.')
-}
